Netmap bindings for libuv
Netmap is a project that gives user-space networking to linux
applications. Netmap binds to the network interface, and traffic
steering rules (generated by ethtool) drive packets into user-specific
receive queues. This completely bypasses the kernel except for NIC IRQs.
Because only some queues are controlled in userspace, the kernel
continues to run other functionality like pings and ssh. In practice we
find this yields approximately 5x more efficient packet receives.

https://blog.cloudflare.com/single-rx-queue-kernel-bypass-with-netmap/

This patch brings netmap to libuv for UDP packets. We connect the event
loop to a single netmap queue. When the `use_netmap` flag is turned on
for new udp sockets, these sockets bypass the kernel calls and instead
operate on top of the netmap rings.