win: allow fatal errors to trigger WER
this is a backport of https://github.com/libuv/libuv/commit/da343e6

our libuv gets linked into the Social SDK, which causes applications
linking to it to not be able to write local crash dumps when the windows
LocalDumps registry feature is enabled. Some game devs use this to
capture crashes during testing, so that's not very nice of us.